refactor and organize into packages
