+++ worker/src/main/java/gov/cms/ab2d/worker/config/JobMessageSource.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.worker;
+package gov.cms.ab2d.worker.config;
 
 import org.springframework.integration.jdbc.JdbcPollingChannelAdapter;
