+++ worker/src/main/resources/application.properties
@@ -5,9 +5,10 @@ spring.datasource.password=${AB2D_DB_PASSWORD}
 spring.datasource.driver-class-name=org.postgresql.Driver
 
 ## ---------------------------------------------------------------------------------------------------  JPA CONFIG
+spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.temp.use_jdbc_metadata_defaults=false
-spring.jpa.hibernate.ddl-auto=validate
+
 
 ## ------------------------------------------------------------------------------------  SPRING INTEGRATION CONFIG
 spring.integration.jdbc.initialize-schema=always
